Split jwtAuth in required or optional
